+++ src/firestore/firestore.module.ts
@@ -1,5 +1,6 @@
-import { InjectionToken, ModuleWithProviders, NgModule } from '@angular/core';
-import { AngularFirestore, EnablePersistenceToken } from './firestore';
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { PersistenceSettings } from './interfaces';
+import { AngularFirestore, EnablePersistenceToken, PersistenceSettingsToken } from './firestore';
 
 import 'firebase/firestore';
 
@@ -10,11 +11,12 @@ export class AngularFirestoreModule {
   /**
    * Attempt to enable persistent storage, if possible
    */
-  static enablePersistence(): ModuleWithProviders {
+  static enablePersistence(persistenceSettings?: PersistenceSettings): ModuleWithProviders {
     return {
       ngModule: AngularFirestoreModule,
       providers: [
         { provide: EnablePersistenceToken, useValue: true },
+        { provide: PersistenceSettingsToken, useValue: persistenceSettings },
       ]
     }
   }